refactor: move tasks/features endpoint --> projects/{id}/task-completion
