+++ src/backend/app/tasks/tasks_routes.py
@@ -21,16 +21,12 @@ from datetime import datetime, timedelta
 from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException
-from loguru import logger as log
 from sqlalchemy.orm import Session
-from sqlalchemy.sql import text
 
 from app.auth.osm import AuthUser, login_required
 from app.auth.roles import get_uid, mapper
-from app.central import central_crud
 from app.db import database
 from app.models.enums import TaskStatus
-from app.projects import project_crud, project_deps
 from app.tasks import tasks_crud, tasks_schemas
 
 router = APIRouter(
@@ -136,74 +132,6 @@ async def update_task_status(
     if not task:
         raise HTTPException(status_code=404, detail="Task status could not be updated.")
     return await tasks_crud.update_task_history(task, db)
-
-
-@router.get("/features/")
-async def task_features_count(
-    project_id: int,
-    db: Session = Depends(database.get_db),
-):
-    """Get all features within a task area."""
-    # Get the project object.
-    project = await project_crud.get_project(db, project_id)
-
-    # ODK Credentials
-    odk_credentials = await project_deps.get_odk_credentials(db, project_id)
-
-    odk_details = central_crud.list_odk_xforms(project.odkid, odk_credentials, True)
-
-    # Assemble the final data list
-    data = []
-    feature_count_query = text(
-        """
-        SELECT id, feature_count
-        FROM tasks
-        WHERE project_id = :project_id;
-    """
-    )
-    result = db.execute(feature_count_query, {"project_id": project_id})
-    feature_counts = result.all()
-
-    if not feature_counts:
-        msg = f"To tasks found for project {project_id}"
-        log.warning(msg)
-        raise HTTPException(status_code=404, detail=msg)
-
-    data.extend(
-        {
-            "task_id": record[0],
-            "submission_count": odk_details[0]["submissions"],
-            "last_submission": odk_details[0]["lastSubmission"],
-            "feature_count": record[1],
-        }
-        for record in feature_counts
-    )
-
-    return data
-
-    # @router.get(
-    # "/task-comments/", response_model=list[tasks_schemas.TaskCommentResponse]
-    # )
-    # async def task_comments(
-    #     project_id: int,
-    #     task_id: int,
-    #     db: Session = Depends(database.get_db),
-    # ):
-    #     """Retrieve a list of task comments for a specific project and task.
-
-    #     Args:
-    #         project_id (int): The ID of the project.
-    #         task_id (int): The ID of the task.
-    #         db (Session, optional): The database session.
-
-    #     Returns:
-    #         A list of task comments.
-    #     """
-    #     task_comment_list = await tasks_crud.get_task_comments(
-    #               db, project_id, task_id
-    #               )
-
-    # return task_comment_list
 
 
 @router.post("/task-comments/", response_model=tasks_schemas.TaskCommentResponse)
